makefile: do not update kernels containing BTF info
Kernels being released for some distros, currently, support BTF
information. It is important to keep BTF files for already released
kernels (that does not support BTF) but there are some BTF files that
could be removed from the HUB (because their kernel started supporting
BTF).

This is true for:

- fedora33 (released kernels support BTF)
- fedora34 (released kernels support BTF)
- bionic (5.4 kernels started supporting BTF)
- focal
  - 5.4 kernels started supporting BTF
  - 5.8 HWE kernels were outdated by 5.11 kernels
  - 5.11 HWE kernels were uotdated by 5.13 kernels
  - 5.13 kernels support BTF
- centos8 (kernels being released support BTF)

This makes the "update" target not to check for newer kernels of those
distributions (as it does not make sense).

Note: More and more BTFHUB will take care of legacy kernels, as designed
from the beginning. Idea is to keep the HUB with BTF files for the
released kernels that don't support BTF (so users that did not upgrade
their environments are able to run CO-RE eBPF programs).