README: preserving source line references (#40)
The README refers to specific source code lines in some links. Not 
using permalinks, though, means that as the references get updated 
the links become stale and point to completely different places in 
the code.

This PR changes those references to permalinks so that the exact line 
is preserved in the reference.