respect system's bpftool path (#110)
Some systems ship with very old versions of `bpftool` and updating using
`apt`; although successful, never really brings the up-to-date version.

One way to tackle this is to build from source. Hard-coding the path of
the bpftool on those systems results in the use of the older binary.

This change, will give the system a chance to use the intended bpftool.